Rename packages, update java app name
