+++ purchase-tester-java/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Purchases Sample Java</string>
+    <string name="app_name">Purchase Tester Java</string>
 </resources>
