+++ purchase-tester/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
         package="com.revenuecat.sample">
 
     <application
-            android:name="com.revenuecat.sample.MainApplication"
+            android:name="com.revenuecat.purchasetester.MainApplication"
             android:allowBackup="true"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
@@ -11,7 +11,7 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
         <activity
-                android:name="com.revenuecat.sample.MainActivity"
+                android:name="com.revenuecat.purchasetester.MainActivity"
                 android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
