+++ purchase-tester/src/main/java/com/revenuecat/purchasetester/Constants.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.sample
+package com.revenuecat.purchasetester
 
 const val JSON_FORMATTER_INDENT_SPACES = 4
 const val ANIMATION_HALF_ROTATION_DEGREES = 180f
