+++ purchase-tester/src/main/java/com/revenuecat/purchasetester/EntitlementInfoExtensions.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.sample
+package com.revenuecat.purchasetester
 
 import com.revenuecat.purchases.EntitlementInfo
