+++ purchase-tester/src/main/java/com/revenuecat/purchasetester/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.revenuecat.sample
+package com.revenuecat.purchasetester
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.revenuecat.sample.R
 
 class MainActivity : AppCompatActivity() {
